+++ lib/twi/error.rb
@@ -1,0 +1,3 @@
+module Twi
+  Error = Class.new StandardError
+end