+++ lib/twi/mock.rb
@@ -1,0 +1,5 @@
+module Twi
+  class Mock
+    attr_accessor :phone, :phone_error
+  end
+end