+++ lib/twi/mock.rb
@@ -1,5 +1,5 @@
 module Twi
   class Mock
-    attr_accessor :phone, :phone_error, :message, :conversation, :conversation_error
+    attr_accessor :phone, :phone_error, :message, :conversation, :conversation_error, :medium
   end
 end