+++ lib/twi/mock/event.rb
@@ -1,0 +1,6 @@
+module Twi
+  class Mock::Event < Event
+  private
+    def medium_for(params) = Mock::Medium.new params
+  end
+end