+++ lib/twi/error.rb
@@ -1,5 +1,11 @@
 module Twi
   class Error < StandardError
+    attr_reader :code
+
+    def initialize(attributes = {})
+      @code = attributes[:code]
+      super attributes[:message]
+    end
   end
 
   class ExistingConversationError < Error