+++ lib/twi/message.rb
@@ -39,7 +39,8 @@ module Twi
 
       @id = message.sid
       @status = message.status
-      # todo rescue and then set @code
+    rescue Twilio::REST::RestError => error
+      raise Error, error
     end
 
     # @return [Hash] the shape of the payload send by Twilio to the callback URL.