+++ lib/twi/mock.rb
@@ -1,5 +1,8 @@
 module Twi
   class Mock
-    attr_accessor :phone, :phone_error, :message, :conversation, :conversation_error, :medium
+    attr_accessor :phone, :phone_error
+    attr_accessor :message, :message_error
+    attr_accessor :conversation, :conversation_error
+    attr_accessor :medium
   end
 end